+++ app/sdks/client-web/types/index.d.ts
@@ -321,9 +321,9 @@ declare namespace Appwrite {
          * @param {number} height
          * @param {number} quality
          * @throws {Error}
-         * @return {Promise}         
+         * @return {string}         
          */
-	    getBrowser(code: string, width: number, height: number, quality: number): Promise<object>;
+	    getBrowser(code: string, width: number, height: number, quality: number): string;
 
         /**
          * Get Credit Card Icon